// public/apps/the_peoples_edh_bracket/the-peoples-edh-bracket-client/src/types/bindings/CardRatingEnriched.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CardRatingReviews } from "./CardRatingReviews";

export type CardRatingEnriched = { total_points: string, global_points: string, reviews: CardRatingReviews, card_name: string, uuid: string, card_oracle_id: string, rater_person_uuid: string, points: string, reason: string | null, created_at: string, updated_at: string | null, };