// public/apps/the_peoples_edh_bracket/the-peoples-edh-bracket-client/src/types/bindings/GetRatingsParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GetRatingsParamsSort } from "./GetRatingsParamsSort";

export type GetRatingsParams = { card_oracle_id: string | null, rater_person_uuid: string | null, q: string | null, sort: GetRatingsParamsSort | null, page: number, page_size: number, };